refactor: RemoteControlPanel 타입 분리
