+++ front/src/type/RemoteControlPanel/Input-type.ts
@@ -1,0 +1,6 @@
+export interface ChatInputProps {
+  handleSubmit: (e: React.FormEvent<HTMLFormElement>) => void;
+  input: string;
+  setInput: React.Dispatch<React.SetStateAction<string>>;
+  isAsking: boolean;
+}
