+++ front/src/type/RemoteControlPanel/Messages-type.ts
@@ -1,0 +1,5 @@
+import { QA } from "../QA-type";
+
+export interface ChatMessagesProps {
+  chats: QA[];
+}
