+++ front/src/type/RemoteControlPanel/MiniMode-type.ts
@@ -1,0 +1,3 @@
+export interface MiniModeProps {
+  onExpand: () => void;
+}
