+++ front/src/type/RemoteControlPanel/SetMiniMode-type.ts
@@ -1,0 +1,3 @@
+export interface SetMiniModeProps {
+  SetMiniMode: React.Dispatch<React.SetStateAction<boolean>>;
+}
